requirements.txt
paho-mqtt==2.1.0
bleak
pillow
watchdog
